+++ src/pynwb/io/file.py
@@ -1,4 +1,3 @@
-from dateutil.parser import parse as dateutil_parse
 import typing
 
 from hdmf.build import ObjectMapper, Builder, GroupBuilder
@@ -7,7 +6,7 @@ from hdmf.utils import docval, get_docval
 from .. import register_map
 from ..file import NWBFile, Subject
 from ..core import ScratchData
-from .utils import get_nwb_version, NO_OVERRIDE
+from .utils import get_nwb_version, parse_date, NO_OVERRIDE
 
 
 @register_map(NWBFile)
@@ -231,7 +230,7 @@ class NWBFileMap(ObjectMapper):
         for user convenience and consistency with how they are written.
         """
         datestr = builder.get('session_start_time').data
-        date = dateutil_parse(datestr)
+        date = parse_date(datestr, "session_start_time")
         return date
 
     @ObjectMapper.constructor_arg('timestamps_reference_time')
@@ -244,7 +243,7 @@ class NWBFileMap(ObjectMapper):
         for user convenience and consistency with how they are written.
         """
         datestr = builder.get('timestamps_reference_time').data
-        date = dateutil_parse(datestr)
+        date = parse_date(datestr, "timestamps_reference_time")
         return date
 
     @ObjectMapper.constructor_arg('file_create_date')
@@ -257,7 +256,7 @@ class NWBFileMap(ObjectMapper):
         for user convenience and consistency with how they are written.
         """
         datestr = builder.get('file_create_date').data
-        dates = list(map(dateutil_parse, datestr))
+        dates = [parse_date(date_string, "file_create_date") for date_string in datestr]
         return dates
 
     @ObjectMapper.constructor_arg('experimenter')
@@ -348,7 +347,7 @@ class SubjectMap(ObjectMapper):
             return
         else:
             datestr = dob_builder.data
-            date = dateutil_parse(datestr)
+            date = parse_date(datestr, "date_of_birth")
             return date
 
     @ObjectMapper.constructor_arg("age__reference")